Refactor wrapper recipes package
