+++ minecraft/1.7/src/main/java/nova/core/wrapper/mc/forge/v17/launcher/NovaMinecraft.java
@@ -44,8 +44,6 @@ import nova.core.wrapper.mc.forge.v17.depmodules.LanguageModule;
 import nova.core.wrapper.mc.forge.v17.depmodules.NetworkModule;
 import nova.core.wrapper.mc.forge.v17.depmodules.SaveModule;
 import nova.core.wrapper.mc.forge.v17.depmodules.TickerModule;
-import nova.core.wrapper.mc.forge.v17.recipes.MinecraftRecipeRegistry;
-import nova.core.wrapper.mc.forge.v17.recipes.RecipeConverter;
 import nova.core.wrapper.mc.forge.v17.wrapper.CategoryConverter;
 import nova.core.wrapper.mc.forge.v17.wrapper.DirectionConverter;
 import nova.core.wrapper.mc.forge.v17.wrapper.assets.AssetConverter;
@@ -57,6 +55,8 @@ import nova.core.wrapper.mc.forge.v17.wrapper.entity.EntityConverter;
 import nova.core.wrapper.mc.forge.v17.wrapper.inventory.InventoryConverter;
 import nova.core.wrapper.mc.forge.v17.wrapper.item.ItemConverter;
 import nova.core.wrapper.mc.forge.v17.wrapper.item.OreDictionaryIntegration;
+import nova.core.wrapper.mc.forge.v17.wrapper.recipes.MinecraftRecipeRegistry;
+import nova.core.wrapper.mc.forge.v17.wrapper.recipes.RecipeConverter;
 import nova.internal.core.Game;
 import nova.internal.core.bootstrap.DependencyInjectionEntryPoint;
 import nova.internal.core.deps.DepDownloader;
