+++ minecraft/1.7/src/main/java/nova/core/wrapper/mc/forge/v17/wrapper/recipes/MinecraftRecipeRegistry.java
@@ -18,7 +18,7 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package nova.core.wrapper.mc.forge.v17.recipes;
+package nova.core.wrapper.mc.forge.v17.wrapper.recipes;
 
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.CraftingManager;
