+++ minecraft/1.7/src/main/java/nova/core/wrapper/mc/forge/v17/wrapper/recipes/forward/ShapelessRecipeBasic.java
@@ -18,33 +18,26 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package nova.core.wrapper.mc.forge.v17.recipes;
+package nova.core.wrapper.mc.forge.v17.wrapper.recipes.forward;
 
 import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.ItemStack;
+import net.minecraft.item.crafting.ShapelessRecipes;
 import net.minecraft.world.World;
-import net.minecraftforge.oredict.ShapedOreRecipe;
-import nova.core.item.Item;
-import nova.core.recipes.crafting.ShapedCraftingRecipe;
-import nova.internal.core.Game;
+import nova.core.recipes.crafting.ShapelessCraftingRecipe;
+import nova.core.wrapper.mc.forge.v17.wrapper.item.ItemConverter;
+import nova.core.wrapper.mc.forge.v17.wrapper.recipes.backward.MCCraftingGrid;
 
-import java.util.Optional;
+import java.util.Arrays;
 
 /**
  * @author Stan
  */
-public class ShapedRecipeOre extends ShapedOreRecipe {
-	private final ShapedCraftingRecipe recipe;
+public class ShapelessRecipeBasic extends ShapelessRecipes {
+	private final ShapelessCraftingRecipe recipe;
 
-	public ShapedRecipeOre(Object[] contents, ShapedCraftingRecipe recipe) {
-		super((ItemStack) Game.natives().toNative(recipe.getExampleOutput().get()), contents);
-
+	public ShapelessRecipeBasic(ItemStack[] ingredients, ShapelessCraftingRecipe recipe) {
+		super(recipe.getExampleOutput().map(ItemConverter.instance()::toNative).orElse(null), Arrays.asList(ingredients));
 		this.recipe = recipe;
 	}
 
@@ -55,10 +48,6 @@ public class ShapedRecipeOre extends ShapedOreRecipe {
 
 	@Override
 	public ItemStack getCraftingResult(InventoryCrafting inventory) {
-		Optional<Item> craftingResult = recipe.getCraftingResult(MCCraftingGrid.get(inventory));
-		if (craftingResult.isPresent()) {
-			return ((ItemStack) Game.natives().toNative(craftingResult.get())).copy();
-		}
-		return null;
+		return recipe.getCraftingResult(MCCraftingGrid.get(inventory)).map(ItemConverter.instance()::toNative).orElse(null);
 	}
 }
