+++ minecraft/1.7/src/main/java/nova/core/wrapper/mc/forge/v17/wrapper/recipes/forward/ShapelessRecipeOre.java
@@ -18,17 +18,15 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package nova.core.wrapper.mc.forge.v17.recipes;
+package nova.core.wrapper.mc.forge.v17.wrapper.recipes.forward;
 
 import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
 import net.minecraftforge.oredict.ShapelessOreRecipe;
-import nova.core.item.Item;
 import nova.core.recipes.crafting.ShapelessCraftingRecipe;
-import nova.internal.core.Game;
-
-import java.util.Optional;
+import nova.core.wrapper.mc.forge.v17.wrapper.item.ItemConverter;
+import nova.core.wrapper.mc.forge.v17.wrapper.recipes.backward.MCCraftingGrid;
 
 /**
  * @author Stan Hebben
@@ -37,8 +35,7 @@ public class ShapelessRecipeOre extends ShapelessOreRecipe {
 	private final ShapelessCraftingRecipe recipe;
 
 	public ShapelessRecipeOre(Object[] ingredients, ShapelessCraftingRecipe recipe) {
-		super((ItemStack) Game.natives().toNative(recipe.getExampleOutput().get()), ingredients);
-
+		super(recipe.getExampleOutput().map(ItemConverter.instance()::toNative).orElse(null), ingredients);
 		this.recipe = recipe;
 	}
 
@@ -49,10 +46,6 @@ public class ShapelessRecipeOre extends ShapelessOreRecipe {
 
 	@Override
 	public ItemStack getCraftingResult(InventoryCrafting inventory) {
-		Optional<Item> craftingResult = recipe.getCraftingResult(MCCraftingGrid.get(inventory));
-		if (craftingResult.isPresent()) {
-			return ((ItemStack) Game.natives().toNative(craftingResult.get())).copy();
-		}
-		return null;
+		return recipe.getCraftingResult(MCCraftingGrid.get(inventory)).map(ItemConverter.instance()::toNative).orElse(null);
 	}
 }
