+++ minecraft/1.8/src/main/java/nova/core/wrapper/mc/forge/v18/launcher/NovaMinecraft.java
@@ -46,8 +46,6 @@ import nova.core.wrapper.mc.forge.v18.depmodules.LanguageModule;
 import nova.core.wrapper.mc.forge.v18.depmodules.NetworkModule;
 import nova.core.wrapper.mc.forge.v18.depmodules.SaveModule;
 import nova.core.wrapper.mc.forge.v18.depmodules.TickerModule;
-import nova.core.wrapper.mc.forge.v18.recipes.MinecraftRecipeRegistry;
-import nova.core.wrapper.mc.forge.v18.recipes.RecipeConverter;
 import nova.core.wrapper.mc.forge.v18.wrapper.CategoryConverter;
 import nova.core.wrapper.mc.forge.v18.wrapper.DirectionConverter;
 import nova.core.wrapper.mc.forge.v18.wrapper.VectorConverter;
@@ -60,6 +58,8 @@ import nova.core.wrapper.mc.forge.v18.wrapper.entity.EntityConverter;
 import nova.core.wrapper.mc.forge.v18.wrapper.inventory.InventoryConverter;
 import nova.core.wrapper.mc.forge.v18.wrapper.item.ItemConverter;
 import nova.core.wrapper.mc.forge.v18.wrapper.item.OreDictionaryIntegration;
+import nova.core.wrapper.mc.forge.v18.wrapper.recipes.MinecraftRecipeRegistry;
+import nova.core.wrapper.mc.forge.v18.wrapper.recipes.RecipeConverter;
 import nova.internal.core.Game;
 import nova.internal.core.bootstrap.DependencyInjectionEntryPoint;
 import nova.internal.core.deps.DepDownloader;
