+++ minecraft/1.8/src/main/java/nova/core/wrapper/mc/forge/v18/wrapper/recipes/forward/ShapedRecipeOre.java
@@ -18,28 +18,30 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package nova.core.wrapper.mc.forge.v17.recipes;
+/*
+ * To change this license header, choose License Headers in Project Properties.
+ * To change this template file, choose Tools | Templates
+ * and open the template in the editor.
+ */
+
+package nova.core.wrapper.mc.forge.v18.wrapper.recipes.forward;
 
 import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.ShapelessRecipes;
 import net.minecraft.world.World;
-import nova.core.item.Item;
-import nova.core.recipes.crafting.ShapelessCraftingRecipe;
-import nova.internal.core.Game;
-
-import java.util.Arrays;
-import java.util.Optional;
+import net.minecraftforge.oredict.ShapedOreRecipe;
+import nova.core.recipes.crafting.ShapedCraftingRecipe;
+import nova.core.wrapper.mc.forge.v18.wrapper.item.ItemConverter;
+import nova.core.wrapper.mc.forge.v18.wrapper.recipes.backward.MCCraftingGrid;
 
 /**
  * @author Stan
  */
-public class ShapelessRecipeBasic extends ShapelessRecipes {
-	private final ShapelessCraftingRecipe recipe;
+public class ShapedRecipeOre extends ShapedOreRecipe {
+	private final ShapedCraftingRecipe recipe;
 
-	public ShapelessRecipeBasic(ItemStack[] ingredients, ShapelessCraftingRecipe recipe) {
-		super(recipe.getExampleOutput().isPresent() ? Game.natives().toNative(recipe.getExampleOutput().get()) : null, Arrays.asList(ingredients));
-
+	public ShapedRecipeOre(Object[] contents, ShapedCraftingRecipe recipe) {
+		super(recipe.getExampleOutput().map(ItemConverter.instance()::toNative).orElse(null), contents);
 		this.recipe = recipe;
 	}
 
@@ -50,10 +52,6 @@ public class ShapelessRecipeBasic extends ShapelessRecipes {
 
 	@Override
 	public ItemStack getCraftingResult(InventoryCrafting inventory) {
-		Optional<Item> craftingResult = recipe.getCraftingResult(MCCraftingGrid.get(inventory));
-		if (craftingResult.isPresent()) {
-			return Game.natives().toNative(craftingResult.get());
-		}
-		return null;
+		return recipe.getCraftingResult(MCCraftingGrid.get(inventory)).map(ItemConverter.instance()::toNative).orElse(null);
 	}
 }
