+++ minecraft/1.8/src/main/java/nova/core/wrapper/mc/forge/v18/wrapper/recipes/forward/ShapelessRecipeOre.java
@@ -18,33 +18,24 @@
  * along with NOVA.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-package nova.core.wrapper.mc.forge.v18.recipes;
+package nova.core.wrapper.mc.forge.v18.wrapper.recipes.forward;
 
 import net.minecraft.inventory.InventoryCrafting;
 import net.minecraft.item.ItemStack;
 import net.minecraft.world.World;
-import net.minecraftforge.oredict.ShapedOreRecipe;
-import nova.core.item.Item;
-import nova.core.recipes.crafting.ShapedCraftingRecipe;
-import nova.internal.core.Game;
-
-import java.util.Optional;
+import net.minecraftforge.oredict.ShapelessOreRecipe;
+import nova.core.recipes.crafting.ShapelessCraftingRecipe;
+import nova.core.wrapper.mc.forge.v18.wrapper.item.ItemConverter;
+import nova.core.wrapper.mc.forge.v18.wrapper.recipes.backward.MCCraftingGrid;
 
 /**
- * @author Stan
+ * @author Stan Hebben
  */
-public class ShapedRecipeOre extends ShapedOreRecipe {
-	private final ShapedCraftingRecipe recipe;
+public class ShapelessRecipeOre extends ShapelessOreRecipe {
+	private final ShapelessCraftingRecipe recipe;
 
-	public ShapedRecipeOre(Object[] contents, ShapedCraftingRecipe recipe) {
-		super((ItemStack) Game.natives().toNative(recipe.getExampleOutput().get()), contents);
-
+	public ShapelessRecipeOre(Object[] ingredients, ShapelessCraftingRecipe recipe) {
+		super(recipe.getExampleOutput().map(ItemConverter.instance()::toNative).orElse(null), ingredients);
 		this.recipe = recipe;
 	}
 
@@ -55,10 +46,6 @@ public class ShapedRecipeOre extends ShapedOreRecipe {
 
 	@Override
 	public ItemStack getCraftingResult(InventoryCrafting inventory) {
-		Optional<Item> craftingResult = recipe.getCraftingResult(MCCraftingGrid.get(inventory));
-		if (craftingResult.isPresent()) {
-			return ((ItemStack) Game.natives().toNative(craftingResult.get())).copy();
-		}
-		return null;
+		return recipe.getCraftingResult(MCCraftingGrid.get(inventory)).map(ItemConverter.instance()::toNative).orElse(null);
 	}
 }
